TW-4638: trim docstrings and consolidate RSVP test files
Cleanup follow-up to eba27dc — reduces over-engineered comments
and removes the "_extra_test.go" smell. No behavior change;
make ci-full clean (180 packages ok, 0 FAIL).

- Merge calendars_events_rsvp_extra_test.go into the canonical
  calendars_events_rsvp_test.go (combined 474 lines, under cap).
- Rename handlers_email_rsvp_extra_test.go ->
  handlers_email_rsvp_edge_cases_test.go (the canonical 542-line
  file is too close to the 600-line cap to merge).
- Trim verbose privacy/security docstrings in handlers_helpers.go
  (parseJSONBody 9->3, writeBadParamError 5->2, writeUpstreamError
  11->3 lines).
- Trim handlers_email_rsvp.go: errEventNotImported / errNoWritable
  Calendar 5-6 -> 2 lines each; handleEmailRSVP pipeline doc 13 -> 5
  lines (kept the security-invariant paragraph); 6 inline comment
  blocks in the body trimmed; findInviteEventAcrossCalendars 17 -> 4
  lines; writableCalendars 3 -> 1 line; findEventByICalUID 6 -> 2
  lines.
- Drop the writeError-vs-http.Error preamble in
  handlers_availability.go and trim parseInt64Param /
  parseIntParam docstrings.
